refactor[#523] : widgets/follow/ui 타입 리팩토링
